+++ test/transform/resource/before/FieldDefaultsViaConfigAndRequiredArgsConstructor.java
@@ -1,0 +1,5 @@
+//CONF: lombok.fieldDefaults.defaultFinal = true
+@lombok.RequiredArgsConstructor
+class FieldDefaultsViaConfigAndRequiredArgsConstructor {
+	int x;
+}